dom/base/test/test_bug421602.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=421602
-->
<head>
  <title>Test for Bug 421602</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=421602">Mozilla Bug 421602</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 421602 **/
SimpleTest.waitForExplicitFinish();

var img1loaded = false;
var img1errored = false;

// Our test image
function loadTestImage() {
  var img1 = new Image();
  img1.onload = function() {
    img1loaded = true;
    finishTest();
  }
  img1.onerror = function() {
    img1errored = true;
    finishTest();
  }
  img1.src = window.location.href + "?image1=true";
}
loadTestImage();

// Probably overkill to gc() more than once, but let's be safe
SpecialPowers.gc(); SpecialPowers.gc(); SpecialPowers.gc();

function finishTest() {
  is(img1errored, true, "Image 1 should error");
  is(img1loaded, false, "Image 1 should not load");
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
